+++ gsa/src/web/entities/container.js
@@ -202,7 +202,7 @@ class EntitiesContainer extends React.Component {
   }
 
   startTimer() {
-    if (!this.isRunning) {
+    if (!this.isRunning || isDefined(this.timer)) {
       return;
     }