+++ gsa/src/web/entity/container.js
@@ -86,7 +86,7 @@ class EntityContainer extends React.Component {
   }
 
   startTimer() {
-    if (!this.isRunning) {
+    if (!this.isRunning || isDefined(this.timer)) {
       return;
     }