+++ CHANGES.md
@@ -8,6 +8,7 @@ $ cd gsa && git checkout gsa-8.0 && git log
 
 ## gsa 8.0.1 (unreleased)
 
+ * Don't run more then one reload timer for a page #1289
  * Set first=1 when starting delta report selection #1288
  * Fix pagination with default filter (reset filter.id if filter is changed) #1288
  * Fix setting filters at report details page with less and greater then relations